+++ code/modules/events/spacevine.dm
@@ -1,29 +1,30 @@
-/datum/round_event_control/spacevine
+/datum/round_event_control/spacevine // rogue version in vines.dm, uses below round_event/
 	name = "Spacevine"
 	typepath = /datum/round_event/spacevine
 	weight = 15
 	max_occurrences = 3
 	min_players = 10
 
-/datum/round_event/spacevine
-	fakeable = FALSE
-
 /datum/round_event/spacevine/start()
 	var/list/turfs = list() //list of all the empty floor turfs in the hallway areas
 
 	var/obj/structure/spacevine/SV = new()
 
-	for(var/area/hallway/A in world)
-		for(var/turf/F in A)
+	for(var/area/rogue/outdoors/town/A in world)
+		for(var/turf/open/F in A)
 			if(F.Enter(SV))
-				turfs += F
+				if(!istype(F, /turf/open/transparent/openspace))
+					turfs += F
 
 	qdel(SV)
 
-	if(turfs.len) //Pick a turf to spawn at if we can
-		var/turf/T = pick(turfs)
-		new /datum/spacevine_controller(T, event = src) //spawn a controller at turf
-
+//	var/maxi = max(GLOB.badomens.len, 1)
+	var/maxi = 7
+	for(var/i in 1 to rand(5,maxi))
+		if(turfs.len) //Pick a turf to spawn at if we can
+			var/turf/T = pick_n_take(turfs)
+			message_admins("VINES at [ADMIN_VERBOSEJMP(T)]")
+			new /datum/spacevine_controller(T, event = src) //spawn a controller at turf
 
 /datum/spacevine_mutation
 	var/name = ""
@@ -92,7 +93,7 @@
 	if(issilicon(crosser))
 		return
 	if(prob(severity) && istype(crosser) && !isvineimmune(crosser))
-		to_chat(crosser, "<span class='alert'>You accidentally touch the vine and feel a strange sensation.</span>")
+		to_chat(crosser, "<span class='alert'>I accidentally touch the vine and feel a strange sensation.</span>")
 		crosser.adjustToxLoss(5)
 
 /datum/spacevine_mutation/toxicity/on_eat(obj/structure/spacevine/holder, mob/living/eater)
@@ -230,13 +231,13 @@
 	if(prob(severity) && istype(crosser) && !isvineimmune(holder))
 		var/mob/living/M = crosser
 		M.adjustBruteLoss(5)
-		to_chat(M, "<span class='alert'>You cut yourself on the thorny vines.</span>")
+		to_chat(M, "<span class='alert'>I cut myself on the thorny vines.</span>")
 
 /datum/spacevine_mutation/thorns/on_hit(obj/structure/spacevine/holder, mob/living/hitter, obj/item/I, expected_damage)
 	if(prob(severity) && istype(hitter) && !isvineimmune(holder))
 		var/mob/living/M = hitter
 		M.adjustBruteLoss(5)
-		to_chat(M, "<span class='alert'>You cut yourself on the thorny vines.</span>")
+		to_chat(M, "<span class='alert'>I cut myself on the thorny vines.</span>")
 	. =	expected_damage
 
 /datum/spacevine_mutation/woodening
@@ -273,8 +274,8 @@
 
 // SPACE VINES (Note that this code is very similar to Biomass code)
 /obj/structure/spacevine
-	name = "space vines"
-	desc = "An extremely expansionistic species of vine."
+	name = "weepvine"
+	desc = ""
 	icon = 'icons/effects/spacevines.dmi'
 	icon_state = "Light1"
 	anchored = TRUE
@@ -282,17 +283,24 @@
 	layer = SPACEVINE_LAYER
 	mouse_opacity = MOUSE_OPACITY_OPAQUE //Clicking anywhere on the turf is good enough
 	pass_flags = PASSTABLE | PASSGRILLE
-	max_integrity = 50
+	max_integrity = 5
+	resistance_flags = FLAMMABLE
+	damage_deflection = 5
+	blade_dulling = DULLING_CUT
 	var/energy = 0
 	var/datum/spacevine_controller/master = null
 	var/list/mutations = list()
+	break_sound = "plantcross"
+	destroy_sound = null
 
 /obj/structure/spacevine/Initialize()
 	. = ..()
+	dir = pick(GLOB.cardinals)
+	icon_state = "Light[rand(1,2)]"
 	add_atom_colour("#ffffff", FIXED_COLOUR_PRIORITY)
 
 /obj/structure/spacevine/examine(mob/user)
-	. = ..()
+	. = ..()/*
 	var/text = "This one is a"
 	if(mutations.len)
 		for(var/A in mutations)
@@ -301,7 +309,7 @@
 	else
 		text += " normal"
 	text += " vine."
-	. += text
+	. += text*/
 
 /obj/structure/spacevine/Destroy()
 	for(var/datum/spacevine_mutation/SM in mutations)
@@ -330,30 +338,35 @@
 		qdel(src)
 
 /obj/structure/spacevine/attacked_by(obj/item/I, mob/living/user)
-	var/damage_dealt = I.force
-	if(I.get_sharpness())
-		damage_dealt *= 4
-	if(I.damtype == BURN)
-		damage_dealt *= 4
+	..()
+//	var/damage_dealt = I.force
+//	if(I.get_sharpness())
+//		damage_dealt *= 4
+//	if(I.damtype == BURN)
+//		damage_dealt *= 4
 
-	for(var/datum/spacevine_mutation/SM in mutations)
-		damage_dealt = SM.on_hit(src, user, I, damage_dealt) //on_hit now takes override damage as arg and returns new value for other mutations to permutate further
-	take_damage(damage_dealt, I.damtype, "melee", 1)
+//	for(var/datum/spacevine_mutation/SM in mutations)
+//		damage_dealt = SM.on_hit(src, user, I, damage_dealt) //on_hit now takes override damage as arg and returns new value for other mutations to permutate further
+//	take_damage(damage_dealt, I.damtype, "melee", 1)
 
 /obj/structure/spacevine/play_attack_sound(damage_amount, damage_type = BRUTE, damage_flag = 0)
 	switch(damage_type)
 		if(BRUTE)
 			if(damage_amount)
-				playsound(src, 'sound/weapons/slash.ogg', 50, TRUE)
+				playsound(src, 'sound/misc/woodhit.ogg', 100, FALSE)
 			else
-				playsound(src, 'sound/weapons/tap.ogg', 50, TRUE)
+				playsound(src, "nodmg", 100, FALSE)
 		if(BURN)
-			playsound(src.loc, 'sound/items/welder.ogg', 100, TRUE)
+			playsound(src.loc, "burn", 100, TRUE)
 
 /obj/structure/spacevine/Crossed(mob/crosser)
 	if(isliving(crosser))
 		for(var/datum/spacevine_mutation/SM in mutations)
 			SM.on_cross(src, crosser)
+	if(prob(23) && istype(crosser) && !isvineimmune(crosser))
+		var/mob/living/M = crosser
+		M.adjustBruteLoss(5)
+		to_chat(M, "<span class='warning'>I nick myself on the thorny vines.</span>")
 
 //ATTACK HAND IGNORING PARENT RETURN VALUE
 /obj/structure/spacevine/attack_hand(mob/user)
@@ -372,9 +385,10 @@
 
 /datum/spacevine_controller
 	var/list/obj/structure/spacevine/vines
+	var/obj/structure/flora/roguetree/evil/tree
 	var/list/growth_queue
-	var/spread_multiplier = 5
-	var/spread_cap = 30
+	var/spread_multiplier = 1
+	var/spread_cap = 4
 	var/list/vine_mutations_list
 	var/mutativeness = 1
 
@@ -389,9 +403,11 @@
 	init_subtypes(/datum/spacevine_mutation/, vine_mutations_list)
 	if(potency != null)
 		mutativeness = potency / 10
-	if(production != null)
-		spread_cap *= production / 5
-		spread_multiplier /= production / 5
+//	if(production != null)
+//		spread_cap *= production / 5
+//		spread_multiplier /= production / 5
+	tree = new /obj/structure/flora/roguetree/evil(location)
+	tree.controller = src
 
 /datum/spacevine_controller/vv_get_dropdown()
 	. = ..()
@@ -423,37 +439,46 @@
 		SV.mutations |= parent.mutations
 		var/parentcolor = parent.atom_colours[FIXED_COLOUR_PRIORITY]
 		SV.add_atom_colour(parentcolor, FIXED_COLOUR_PRIORITY)
-		if(prob(mutativeness))
-			var/datum/spacevine_mutation/randmut = pick(vine_mutations_list - SV.mutations)
-			randmut.add_mutation_to_vinepiece(SV)
+//		if(prob(mutativeness))
+//			var/datum/spacevine_mutation/randmut = pick(vine_mutations_list - SV.mutations)
+//			randmut.add_mutation_to_vinepiece(SV)
 
 	for(var/datum/spacevine_mutation/SM in SV.mutations)
 		SM.on_birth(SV)
 	location.Entered(SV)
 	return SV
 
+/datum/spacevine_controller/proc/endvines()
+	for(var/obj/structure/spacevine/V in vines)
+		V.dieepic()
+	qdel(src)
+
 /datum/spacevine_controller/proc/VineDestroyed(obj/structure/spacevine/S)
 	S.master = null
 	vines -= S
 	growth_queue -= S
 	if(!vines.len)
-		var/obj/item/seeds/kudzu/KZ = new(S.loc)
-		KZ.mutations |= S.mutations
-		KZ.set_potency(mutativeness * 10)
-		KZ.set_production((spread_cap / initial(spread_cap)) * 5)
+//		var/obj/item/seeds/kudzu/KZ = new(S.loc)
+//		KZ.mutations |= S.mutations
+//		KZ.set_potency(mutativeness * 10)
+//		KZ.set_production((spread_cap / initial(spread_cap)) * 5)
 		qdel(src)
 
 /datum/spacevine_controller/process()
 	if(!LAZYLEN(vines))
-		qdel(src) //space vines exterminated. Remove the controller
-		return
+		if(!tree)
+			qdel(src) //space vines exterminated. Remove the controller
+			return
+		else
+			spawn_spacevine_piece(tree.loc, null)
 	if(!growth_queue)
 		qdel(src) //Sanity check
 		return
 
 	var/length = 0
 
-	length = min( spread_cap , max( 1 , vines.len / spread_multiplier ) )
+//	length = min( spread_cap , max( 1 , vines.len / spread_multiplier ) )
+	length = 20
 	var/i = 0
 	var/list/obj/structure/spacevine/queue_end = list()
 
@@ -470,21 +495,29 @@
 				SV.grow()
 		else //If tile is fully grown
 			SV.entangle_mob()
-
+		if(vines.len > 25)
+			break
 		SV.spread()
 		if(i >= length)
 			break
 
 	growth_queue = growth_queue + queue_end
 
+/obj/structure/spacevine/proc/dieepic()
+	icon_state = "[icon_state]d"
+	max_integrity = 1
+	obj_integrity = 1
+	destroy_sound = 'sound/foley/breaksound.ogg'
+
 /obj/structure/spacevine/proc/grow()
 	if(!energy)
-		src.icon_state = pick("Med1", "Med2", "Med3")
+		src.icon_state = pick("Med1", "Med2")
 		energy = 1
 		set_opacity(1)
 	else
-		src.icon_state = pick("Hvy1", "Hvy2", "Hvy3")
+		src.icon_state = pick("Hvy1", "Hvy2")
 		energy = 2
+//		density = TRUE
 
 	for(var/datum/spacevine_mutation/SM in mutations)
 		SM.on_grow(src)
@@ -503,8 +536,9 @@
 	for(var/datum/spacevine_mutation/SM in mutations)
 		SM.on_buckle(src, V)
 	if((V.stat != DEAD) && (V.buckled != src)) //not dead or captured
-		to_chat(V, "<span class='danger'>The vines [pick("wind", "tangle", "tighten")] around you!</span>")
+		to_chat(V, "<span class='danger'>The vines [pick("wind", "tangle", "tighten")] around me!</span>")
 		buckle_mob(V, 1)
+	V.adjustOxyLoss(10)
 
 /obj/structure/spacevine/proc/spread()
 	var/direction = pick(GLOB.cardinals)
@@ -523,7 +557,7 @@
 	var/i
 	for(var/datum/spacevine_mutation/SM in mutations)
 		i += SM.on_explosion(severity, target, src)
-	if(!i && prob(100/severity))
+	if(!i)
 		qdel(src)
 
 /obj/structure/spacevine/temperature_expose(null, temp, volume)
